feat: #8 home 화면 조회

+++ src/main/java/com/example/dogcument/domain/home/dto/HomeResDto.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class HomeResDto {
-	private Integer totalDogs;
+	private Long totalDogs;
 	private LocalDateTime nextFeedingTime;
 
 	private Double neuteredRate;
